Extract diff collection, prompt fingerprint, and mention write reply slices.
Move collectDiffContext, buildReviewPromptFingerprint, and issue-write reply helpers into focused modules with re-exports from review.ts; fix missing buildOrderedReviewPhaseSummary import.
