+++ src/handlers/review.ts
@@ -75,7 +75,7 @@ import { evaluateFeedbackSuppressions, adjustConfidenceForFeedback } from "../fe
 import type { SuggestionClusterStore } from "../knowledge/suggestion-cluster-store.ts";
 import { applyClusterScoringWithDegradation } from "../knowledge/suggestion-cluster-degradation.ts";
 import { classifyError, formatErrorComment, postOrUpdateErrorComment } from "../lib/errors.ts";
-import { fetchAllPullRequestFiles, type PullRequestFileMetadata } from "../lib/github-pr-files.ts";
+import { fetchAllPullRequestFiles } from "../lib/github-pr-files.ts";
 import { estimateTimeoutRisk, computeLanguageComplexity } from "../lib/timeout-estimator.ts";
 import {
   ensureReviewBoundednessDisclosureInSummary,
@@ -149,7 +149,7 @@ import {
   type RepoDoctrineProjection,
 } from "../repo-doctrine/contracts.ts";
 import { $ } from "bun";
-import { fetchAndCheckoutPullRequestHeadRef, buildAuthFetchUrl, fetchRemoteTrackingBranch } from "../jobs/workspace.ts";
+import { fetchAndCheckoutPullRequestHeadRef, fetchRemoteTrackingBranch } from "../jobs/workspace.ts";
 import {
   buildReviewFamilyKey,
   createReviewWorkCoordinator,
@@ -218,6 +218,7 @@ import { classifyReviewTimeoutOutcome } from "../review-orchestration/review-tim
 import { logReviewTimeoutClassification } from "../review-orchestration/review-timeout-classification-log.ts";
 import {
   buildExecutorUnavailablePhases,
+  buildOrderedReviewPhaseSummary,
   buildQueueWaitPhase,
   buildReviewDetailsPhaseTimingSummary,
   buildUnavailableReviewPhase,
@@ -258,6 +259,17 @@ import {
   buildShadowSpecialistLogFields,
   buildCandidateVerificationPublicationEvidenceLogFields,
 } from "../review-orchestration/review-specialist-publication-log.ts";
+import {
+  buildReviewPromptFingerprint,
+  type ReviewPromptBuildContext,
+  type ReviewPromptFingerprintResult,
+} from "../review-orchestration/review-prompt-fingerprint.ts";
+export { buildReviewPromptFingerprint, type ReviewPromptBuildContext, type ReviewPromptFingerprintResult } from "../review-orchestration/review-prompt-fingerprint.ts";
+import {
+  collectDiffContext,
+  REVIEW_WORKSPACE_FETCH_DEPTH,
+} from "../review-orchestration/review-diff-collection.ts";
+export { collectDiffContext, REVIEW_WORKSPACE_FETCH_DEPTH } from "../review-orchestration/review-diff-collection.ts";
 import {
   toProductionLogBudgetReasoning,
   toProductionLogCandidateFindingSnapshot,
@@ -368,175 +380,6 @@ type RetrievalContextForPrompt = {
 };
 
 
-type ReviewPromptBuildContext = Parameters<typeof buildReviewPromptDetails>[0];
-
-export type ReviewPromptFingerprintResult = {
-  fingerprint: string | null;
-  missingSignals: string[];
-};
-
-function sha256Hex(input: string): string {
-  return createHash("sha256").update(input).digest("hex");
-}
-
-function hashPromptString(value: string | null | undefined): string | null {
-  if (typeof value !== "string") {
-    return null;
-  }
-
-  return sha256Hex(value);
-}
-
-
-function normalizePromptStringList(values: string[] | undefined, signal: string): { values: string[] | null; missingSignals: string[] } {
-  if (!values) {
-    return { values: [], missingSignals: [] };
-  }
-
-  const normalized = values
-    .map((value) => typeof value === "string" ? value.trim() : "")
-    .filter((value) => value.length > 0);
-
-  if (values.length > 0 && normalized.length !== values.length) {
-    return { values: null, missingSignals: [signal] };
-  }
-
-  return {
-    values: Array.from(new Set(normalized)).sort((a, b) => a.localeCompare(b)),
-    missingSignals: [],
-  };
-}
-
-function summarizeRetrievalContextFingerprint(
-  retrievalContext: ReviewPromptBuildContext["retrievalContext"],
-): { value: Record<string, unknown> | null; missingSignals: string[] } {
-  if (!retrievalContext) {
-    return { value: null, missingSignals: [] };
-  }
-
-  const summarizedFindings: Array<Record<string, unknown>> = [];
-  for (const finding of retrievalContext.findings) {
-    if (
-      typeof finding.findingText !== "string"
-      || typeof finding.severity !== "string"
-      || typeof finding.category !== "string"
-      || typeof finding.path !== "string"
-      || typeof finding.outcome !== "string"
-      || typeof finding.sourceRepo !== "string"
-      || !Number.isFinite(finding.distance)
-    ) {
-      return { value: null, missingSignals: ["retrieval-fingerprint-data"] };
-    }
-
-    summarizedFindings.push({
-      findingTextHash: sha256Hex(finding.findingText),
-      severity: finding.severity,
-      category: finding.category,
-      path: finding.path,
-      line: finding.line ?? null,
-      snippetHash: hashPromptString(finding.snippet),
-      outcome: finding.outcome,
-      distance: Number(finding.distance.toFixed(6)),
-      sourceRepo: finding.sourceRepo.toLowerCase(),
-    });
-  }
-
-  return {
-    value: {
-      maxChars: retrievalContext.maxChars ?? null,
-      findings: summarizedFindings,
-    },
-    missingSignals: [],
-  };
-}
-
-export function buildReviewPromptFingerprint(
-  context: ReviewPromptBuildContext,
-): ReviewPromptFingerprintResult {
-  const missingSignals: string[] = [];
-
-  const owner = context.owner.trim().toLowerCase();
-  const repo = context.repo.trim().toLowerCase();
-  const baseBranch = context.baseBranch.trim();
-  const headBranch = context.headBranch.trim();
-  const prAuthor = context.prAuthor.trim();
-  const normalizedChangedFiles = normalizePromptStringList(context.changedFiles, "changed-files");
-  const focusAreas = normalizePromptStringList(context.focusAreas, "focus-areas");
-  const ignoredAreas = normalizePromptStringList(context.ignoredAreas, "ignored-areas");
-  const prLabels = normalizePromptStringList(context.prLabels, "pr-labels");
-  const focusHints = normalizePromptStringList(context.focusHints, "focus-hints");
-  const retrievalSummary = summarizeRetrievalContextFingerprint(context.retrievalContext);
-
-  missingSignals.push(...normalizedChangedFiles.missingSignals, ...focusAreas.missingSignals, ...ignoredAreas.missingSignals, ...prLabels.missingSignals, ...focusHints.missingSignals, ...retrievalSummary.missingSignals);
-
-  if (!owner || !repo) missingSignals.push("repo-identity");
-  if (!Number.isInteger(context.prNumber) || context.prNumber <= 0) missingSignals.push("pr-number");
-  if (!baseBranch || !headBranch) missingSignals.push("pr-refs");
-  if (normalizedChangedFiles.values !== null && normalizedChangedFiles.values.length === 0) missingSignals.push("changed-files");
-  if (typeof context.prTitle !== "string") missingSignals.push("pr-title");
-  if (!prAuthor) missingSignals.push("pr-author");
-
-  if (missingSignals.length > 0) {
-    return { fingerprint: null, missingSignals: Array.from(new Set(missingSignals)) };
-  }
-
-  const fingerprintPayload = {
-    version: 1,
-    repo: `${owner}/${repo}`,
-    prNumber: context.prNumber,
-    prTitleHash: sha256Hex(context.prTitle),
-    prBodyHash: hashPromptString(context.prBody),
-    prAuthor,
-    baseBranch,
-    headBranch,
-    changedFiles: normalizedChangedFiles.values,
-    customInstructionsHash: hashPromptString(context.customInstructions),
-    checkpointEnabled: context.checkpointEnabled ?? false,
-    mode: context.mode ?? "standard",
-    severityMinLevel: context.severityMinLevel ?? "minor",
-    focusAreas: focusAreas.values,
-    ignoredAreas: ignoredAreas.values,
-    maxComments: context.maxComments ?? null,
-    suppressionsHash: hashPromptString(JSON.stringify(context.suppressions ?? [])),
-    minConfidence: context.minConfidence ?? null,
-    diffAnalysisHash: hashPromptString(JSON.stringify(context.diffAnalysis ?? null)),
-    matchedPathInstructionsHash: hashPromptString(JSON.stringify(context.matchedPathInstructions ?? [])),
-    incrementalContextHash: hashPromptString(JSON.stringify(context.incrementalContext ?? null)),
-    retrievalContext: retrievalSummary.value,
-    reviewPrecedentsHash: hashPromptString(JSON.stringify(context.reviewPrecedents ?? [])),
-    wikiKnowledgeHash: hashPromptString(JSON.stringify(context.wikiKnowledge ?? [])),
-    filesByLanguageHash: hashPromptString(JSON.stringify(context.filesByLanguage ?? null)),
-    outputLanguage: context.outputLanguage ?? null,
-    prLabels: prLabels.values,
-    focusHints: focusHints.values,
-    conventionalTypeHash: hashPromptString(JSON.stringify(context.conventionalType ?? null)),
-    deltaContextHash: hashPromptString(JSON.stringify(context.deltaContext ?? null)),
-    largePRContextHash: hashPromptString(JSON.stringify(context.largePRContext ?? null)),
-    authorTier: context.authorTier ?? null,
-    contributorExperienceContractHash: hashPromptString(JSON.stringify(context.contributorExperienceContract ?? null)),
-    authorExpertiseHash: hashPromptString(JSON.stringify(context.authorExpertise ?? [])),
-    depBumpContextHash: hashPromptString(JSON.stringify(context.depBumpContext ?? null)),
-    searchRateLimitDegradationHash: hashPromptString(JSON.stringify(context.searchRateLimitDegradation ?? null)),
-    isDraft: context.isDraft ?? false,
-    unifiedResultsHash: hashPromptString(JSON.stringify(context.unifiedResults ?? [])),
-    contextWindowHash: hashPromptString(context.contextWindow),
-    clusterPatternsHash: hashPromptString(JSON.stringify(context.clusterPatterns ?? [])),
-    linkedIssuesHash: hashPromptString(JSON.stringify(context.linkedIssues ?? null)),
-    activeRulesHash: hashPromptString(JSON.stringify(context.activeRules ?? [])),
-    graphBlastRadiusHash: hashPromptString(JSON.stringify(context.graphBlastRadius ?? null)),
-    graphContextOptionsHash: hashPromptString(JSON.stringify(context.graphContextOptions ?? null)),
-    structuralImpactHash: hashPromptString(JSON.stringify(context.structuralImpact ?? null)),
-    reviewBoundednessHash: hashPromptString(JSON.stringify(context.reviewBoundedness ?? null)),
-    publishToolNamesHash: hashPromptString(JSON.stringify(context.publishToolNames ?? [])),
-    candidateFindingToolName: context.candidateFindingToolName ?? null,
-    candidateFindingMode: context.candidateFindingMode ?? null,
-  };
-
-  return {
-    fingerprint: sha256Hex(JSON.stringify(fingerprintPayload)),
-    missingSignals: [],
-  };
-}
 
 
 async function fetchCommitMessages(
@@ -570,509 +413,6 @@ async function fetchCommitMessages(
 
 
 
-
-
-type DiffCollectionStrategy =
-  | "triple-dot"
-  | "deepened-triple-dot"
-  | "fallback-two-dot"
-  | "github-file-list-fallback"
-  | "github-pr-files-fallback";
-
-type DiffCollectionResult = {
-  changedFiles: string[];
-  numstatLines: string[];
-  diffContent?: string;
-  strategy: DiffCollectionStrategy;
-  mergeBaseRecovered: boolean;
-  deepenAttempts: number;
-  unshallowAttempted: boolean;
-  diffRange: string;
-};
-
-type DiffCommandResult = {
-  exitCode: number;
-  stdout: string;
-  stderr: string;
-  timedOut: boolean;
-};
-
-type DiffCommandRunner = (args: string[], timeoutMs: number) => Promise<DiffCommandResult>;
-
-type DiffFallbackFile = PullRequestFileMetadata;
-
-export const REVIEW_WORKSPACE_FETCH_DEPTH = 50;
-const DIFF_DEEPEN_STEPS = [50, 150, 300];
-const DIFF_COMMAND_TIMEOUT_MS = 30_000;
-
-async function hasMergeBase(workspaceDir: string, baseRef: string): Promise<boolean> {
-  const mergeBaseResult = await $`git -C ${workspaceDir} merge-base origin/${baseRef} HEAD`.quiet().nothrow();
-  return mergeBaseResult.exitCode === 0;
-}
-
-async function runDiffCommandWithTimeout(params: {
-  workspaceDir: string;
-  args: string[];
-  timeoutMs: number;
-}): Promise<DiffCommandResult> {
-  const { workspaceDir, args, timeoutMs } = params;
-  const proc = Bun.spawn(["git", "-C", workspaceDir, ...args], {
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-
-  const stdoutPromise = new Response(proc.stdout).text();
-  const stderrPromise = new Response(proc.stderr).text();
-  let timer: ReturnType<typeof setTimeout> | undefined;
-  let timedOut = false;
-
-  try {
-    const exitCode = timeoutMs > 0 && Number.isFinite(timeoutMs)
-      ? await Promise.race([
-          proc.exited,
-          new Promise<number>((resolve) => {
-            timer = setTimeout(() => {
-              timedOut = true;
-              try {
-                proc.kill();
-              } catch {
-                // Ignore kill races; the process may have already exited.
-              }
-              resolve(124);
-            }, timeoutMs);
-          }),
-        ])
-      : await proc.exited;
-
-    const [stdout, stderr] = await Promise.all([stdoutPromise, stderrPromise]);
-    return {
-      exitCode,
-      stdout,
-      stderr,
-      timedOut,
-    };
-  } finally {
-    if (timer !== undefined) {
-      clearTimeout(timer);
-    }
-  }
-}
-
-function buildFallbackPatchDiff(files: DiffFallbackFile[]): string | undefined {
-  const chunks = files
-    .filter((file) => typeof file.patch === "string" && file.patch.trim().length > 0)
-    .map((file) => {
-      const oldPath = file.status === "added" ? "/dev/null" : `a/${file.previousFilename ?? file.filename}`;
-      const newPath = file.status === "removed" ? "/dev/null" : `b/${file.filename}`;
-      return [
-        `diff --git a/${file.previousFilename ?? file.filename} b/${file.filename}`,
-        `--- ${oldPath}`,
-        `+++ ${newPath}`,
-        file.patch!.trimEnd(),
-      ].join("\n");
-    });
-
-  return chunks.length > 0 ? chunks.join("\n") + "\n" : undefined;
-}
-
-function buildFallbackNumstatLines(files: DiffFallbackFile[]): string[] {
-  return files.map((file) => {
-    const additions = typeof file.additions === "number" && Number.isFinite(file.additions) ? String(file.additions) : "-";
-    const deletions = typeof file.deletions === "number" && Number.isFinite(file.deletions) ? String(file.deletions) : "-";
-    return `${additions}\t${deletions}\t${file.filename}`;
-  });
-}
-
-async function buildDiffCollectionFallback(params: {
-  fallbackFileProvider?: () => Promise<string[]>;
-  fallbackDiffProvider?: () => Promise<DiffFallbackFile[]>;
-  logger: Logger;
-  baseLog: Record<string, unknown>;
-  stage: string;
-  reason: string;
-  deepenAttempts: number;
-  unshallowAttempted: boolean;
-  mergeBaseRecovered: boolean;
-  diffRange: string;
-}): Promise<DiffCollectionResult> {
-  const {
-    fallbackFileProvider,
-    fallbackDiffProvider,
-    logger,
-    baseLog,
-    stage,
-    reason,
-    deepenAttempts,
-    unshallowAttempted,
-    mergeBaseRecovered,
-    diffRange,
-  } = params;
-
-  if (fallbackDiffProvider) {
-    const fallbackFiles = await fallbackDiffProvider();
-    const uniqueFiles = Array.from(new Map(fallbackFiles.map((file) => [file.filename, file])).values());
-    const changedFiles = uniqueFiles.map((file) => file.filename);
-    const numstatLines = buildFallbackNumstatLines(uniqueFiles);
-    const diffContent = buildFallbackPatchDiff(uniqueFiles);
-    const patchFilesCount = uniqueFiles.filter((file) => typeof file.patch === "string" && file.patch.trim().length > 0).length;
-    const hasCompletePatchFallback = uniqueFiles.length > 0 && patchFilesCount === uniqueFiles.length && diffContent !== undefined;
-    const logFallback = hasCompletePatchFallback ? logger.info.bind(logger) : logger.warn.bind(logger);
-
-    logFallback(
-      {
-        ...baseLog,
-        gate: "diff-collection",
-        stage,
-        reason,
-        strategy: "github-pr-files-fallback",
-        fallbackEvidenceQuality: hasCompletePatchFallback ? "patch-complete" : "patch-partial",
-        deepenAttempts,
-        unshallowAttempted,
-        mergeBaseRecovered,
-        diffRange,
-        changedFilesCount: changedFiles.length,
-        patchFilesCount,
-        diffContentAvailable: diffContent !== undefined,
-      },
-      hasCompletePatchFallback
-        ? "Diff collection used GitHub PR files fallback with patch evidence"
-        : "Diff collection degraded to GitHub PR files fallback",
-    );
-
-    return {
-      changedFiles,
-      numstatLines,
-      diffContent,
-      strategy: "github-pr-files-fallback",
-      mergeBaseRecovered,
-      deepenAttempts,
-      unshallowAttempted,
-      diffRange: "github-api:pr-files",
-    };
-  }
-
-  if (!fallbackFileProvider) {
-    throw new Error(`Diff collection timed out during ${stage} (${reason}) and no fallback provider was configured`);
-  }
-
-  const changedFiles = Array.from(new Set(await fallbackFileProvider()));
-  const boundedFilenameOnlyFallback = changedFiles.length <= 10;
-  const logFallback = boundedFilenameOnlyFallback ? logger.info.bind(logger) : logger.warn.bind(logger);
-
-  logFallback(
-    {
-      ...baseLog,
-      gate: "diff-collection",
-      stage,
-      reason,
-      strategy: "github-file-list-fallback",
-      fallbackEvidenceQuality: boundedFilenameOnlyFallback ? "filename-only-small" : "filename-only",
-      deepenAttempts,
-      unshallowAttempted,
-      mergeBaseRecovered,
-      diffRange,
-      changedFilesCount: changedFiles.length,
-    },
-    "Diff collection degraded to GitHub file-list fallback",
-  );
-
-  return {
-    changedFiles,
-    numstatLines: [],
-    diffContent: undefined,
-    strategy: "github-file-list-fallback",
-    mergeBaseRecovered,
-    deepenAttempts,
-    unshallowAttempted,
-    diffRange: "github-api:file-list",
-  };
-}
-
-export async function collectDiffContext(params: {
-  workspaceDir: string;
-  baseRef: string;
-  maxFilesForFullDiff: number;
-  logger: Logger;
-  baseLog: Record<string, unknown>;
-  token?: string;
-  runGitCommand?: DiffCommandRunner;
-  fallbackFileProvider?: () => Promise<string[]>;
-  fallbackDiffProvider?: () => Promise<DiffFallbackFile[]>;
-  commandTimeoutMs?: number;
-}): Promise<DiffCollectionResult> {
-  const {
-    workspaceDir,
-    baseRef,
-    maxFilesForFullDiff,
-    logger,
-    baseLog,
-    token,
-    runGitCommand,
-    fallbackFileProvider,
-    fallbackDiffProvider,
-    commandTimeoutMs = DIFF_COMMAND_TIMEOUT_MS,
-  } = params;
-
-  const diffCommandRunner = runGitCommand
-    ?? ((args: string[], timeoutMs: number) =>
-      runDiffCommandWithTimeout({ workspaceDir, args, timeoutMs }));
-
-  let strategy: DiffCollectionStrategy = "triple-dot";
-  let mergeBaseRecovered = false;
-  let deepenAttempts = 0;
-  let unshallowAttempted = false;
-
-  // Build auth-injected remote URL once for all fetch calls in this function.
-  const fetchRemote = await buildAuthFetchUrl(workspaceDir, token);
-
-  let mergeBaseAvailable = await hasMergeBase(workspaceDir, baseRef);
-  if (!mergeBaseAvailable) {
-    logger.info(
-      {
-        ...baseLog,
-        gate: "diff-collection",
-        stage: "merge-base-recovery",
-        baseRef,
-        ...toProductionLogRuntimeBudgetFields(commandTimeoutMs),
-      },
-      "Merge base missing before diff collection; attempting history recovery",
-    );
-
-    for (const step of DIFF_DEEPEN_STEPS) {
-      deepenAttempts += 1;
-      logger.info(
-        {
-          ...baseLog,
-          gate: "diff-collection",
-          stage: "merge-base-recovery",
-          attempt: deepenAttempts,
-          deepenBy: step,
-          ...toProductionLogRuntimeBudgetFields(commandTimeoutMs),
-        },
-        "Attempting diff collection merge-base recovery",
-      );
-
-      const deepenResult = await diffCommandRunner(
-        ["fetch", fetchRemote, `+${baseRef}:refs/remotes/origin/${baseRef}`, `--deepen=${step}`],
-        commandTimeoutMs,
-      );
-      if (deepenResult.timedOut) {
-        return await buildDiffCollectionFallback({
-          fallbackFileProvider,
-          fallbackDiffProvider,
-          logger,
-          baseLog,
-          stage: "merge-base-recovery",
-          reason: `fetch-timeout-deepen-${step}`,
-          deepenAttempts,
-          unshallowAttempted,
-          mergeBaseRecovered,
-          diffRange: `origin/${baseRef}...HEAD`,
-        });
-      }
-
-      mergeBaseAvailable = await hasMergeBase(workspaceDir, baseRef);
-      if (mergeBaseAvailable) {
-        mergeBaseRecovered = true;
-        strategy = "deepened-triple-dot";
-        break;
-      }
-    }
-
-    if (!mergeBaseAvailable) {
-      unshallowAttempted = true;
-      logger.info(
-        {
-          ...baseLog,
-          gate: "diff-collection",
-          stage: "merge-base-recovery",
-          attempt: deepenAttempts + 1,
-          mode: "unshallow",
-          ...toProductionLogRuntimeBudgetFields(commandTimeoutMs),
-        },
-        "Attempting diff collection full-history recovery",
-      );
-
-      const unshallowResult = await diffCommandRunner(
-        ["fetch", fetchRemote, `+${baseRef}:refs/remotes/origin/${baseRef}`, "--unshallow"],
-        commandTimeoutMs,
-      );
-      if (unshallowResult.timedOut) {
-        return await buildDiffCollectionFallback({
-          fallbackFileProvider,
-          fallbackDiffProvider,
-          logger,
-          baseLog,
-          stage: "merge-base-recovery",
-          reason: "fetch-timeout-unshallow",
-          deepenAttempts,
-          unshallowAttempted,
-          mergeBaseRecovered,
-          diffRange: `origin/${baseRef}...HEAD`,
-        });
-      }
-
-      mergeBaseAvailable = await hasMergeBase(workspaceDir, baseRef);
-      if (mergeBaseAvailable) {
-        mergeBaseRecovered = true;
-        strategy = "deepened-triple-dot";
-      }
-    }
-  }
-
-  let diffRange = mergeBaseAvailable ? `origin/${baseRef}...HEAD` : `origin/${baseRef}..HEAD`;
-  if (!mergeBaseAvailable) {
-    strategy = "fallback-two-dot";
-  }
-
-  let nameOnlyResult = await diffCommandRunner(["diff", diffRange, "--name-only"], commandTimeoutMs);
-  if (nameOnlyResult.timedOut) {
-    return await buildDiffCollectionFallback({
-      fallbackFileProvider,
-      fallbackDiffProvider,
-      logger,
-      baseLog,
-      stage: "name-only",
-      reason: `diff-timeout-${diffRange}-name-only`,
-      deepenAttempts,
-      unshallowAttempted,
-      mergeBaseRecovered,
-      diffRange,
-    });
-  }
-
-  if (nameOnlyResult.exitCode !== 0 && diffRange.includes("...")) {
-    strategy = "fallback-two-dot";
-    diffRange = `origin/${baseRef}..HEAD`;
-    logger.warn(
-      {
-        ...baseLog,
-        gate: "diff-collection",
-        strategy,
-        reason: "triple-dot-diff-failed",
-      },
-      "Triple-dot diff failed; retrying with deterministic fallback range",
-    );
-    nameOnlyResult = await diffCommandRunner(["diff", diffRange, "--name-only"], commandTimeoutMs);
-    if (nameOnlyResult.timedOut) {
-      return await buildDiffCollectionFallback({
-        fallbackFileProvider,
-        fallbackDiffProvider,
-        logger,
-        baseLog,
-        stage: "name-only",
-        reason: `diff-timeout-${diffRange}-name-only`,
-        deepenAttempts,
-        unshallowAttempted,
-        mergeBaseRecovered,
-        diffRange,
-      });
-    }
-    if (nameOnlyResult.exitCode !== 0) {
-      return await buildDiffCollectionFallback({
-        fallbackFileProvider,
-        fallbackDiffProvider,
-        logger,
-        baseLog,
-        stage: "name-only",
-        reason: `diff-failed-${diffRange}-name-only`,
-        deepenAttempts,
-        unshallowAttempted,
-        mergeBaseRecovered,
-        diffRange,
-      });
-    }
-  } else if (nameOnlyResult.exitCode !== 0) {
-    throw new Error(`git diff ${diffRange} --name-only failed with exit code ${nameOnlyResult.exitCode}`);
-  }
-
-  const changedFiles = splitGitLines(nameOnlyResult.stdout);
-
-  const numstatOutput = await diffCommandRunner(["diff", diffRange, "--numstat"], commandTimeoutMs);
-  let numstatLines: string[] = [];
-  if (numstatOutput.timedOut) {
-    logger.warn(
-      {
-        ...baseLog,
-        gate: "diff-collection",
-        stage: "numstat",
-        diffRange,
-        ...toProductionLogRuntimeBudgetFields(commandTimeoutMs),
-      },
-      "Diff numstat collection timed out; continuing without numstat",
-    );
-  } else if (numstatOutput.exitCode !== 0) {
-    logger.warn(
-      {
-        ...baseLog,
-        gate: "diff-collection",
-        stage: "numstat",
-        diffRange,
-        exitCode: numstatOutput.exitCode,
-      },
-      "Diff numstat collection failed; continuing without numstat",
-    );
-  } else {
-    numstatLines = splitGitLines(numstatOutput.stdout);
-  }
-
-  let diffContent: string | undefined;
-  if (changedFiles.length <= maxFilesForFullDiff) {
-    const fullDiff = await diffCommandRunner(["diff", diffRange], commandTimeoutMs);
-    if (fullDiff.timedOut) {
-      logger.warn(
-        {
-          ...baseLog,
-          gate: "diff-collection",
-          stage: "full-diff",
-          diffRange,
-          changedFilesCount: changedFiles.length,
-          ...toProductionLogRuntimeBudgetFields(commandTimeoutMs),
-        },
-        "Full diff collection timed out; continuing without full diff",
-      );
-    } else if (fullDiff.exitCode !== 0) {
-      logger.warn(
-        {
-          ...baseLog,
-          gate: "diff-collection",
-          stage: "full-diff",
-          diffRange,
-          changedFilesCount: changedFiles.length,
-          exitCode: fullDiff.exitCode,
-        },
-        "Full diff collection failed; continuing without full diff",
-      );
-    } else {
-      diffContent = fullDiff.stdout;
-    }
-  }
-
-  logger.info(
-    {
-      ...baseLog,
-      gate: "diff-collection",
-      strategy,
-      deepenAttempts,
-      unshallowAttempted,
-      mergeBaseRecovered,
-      diffRange,
-      changedFilesCount: changedFiles.length,
-    },
-    "Collected diff context for review",
-  );
-
-  return {
-    changedFiles,
-    numstatLines,
-    diffContent,
-    strategy,
-    mergeBaseRecovered,
-    deepenAttempts,
-    unshallowAttempted,
-    diffRange,
-  };
-}
